requirements.txt
chromadb>=0.5.5
python-docx>=1.1.2
python-dotenv>=1.0.1
openai>=1.40.0
google-genai>=1.30.0
streamlit>=1.39.0
pydantic>=2.8.2
sentence-transformers>=3.0.1
huggingface-hub>=0.24.0
